Feature/routing

+++ src/main/java/dk/itu/drawing/models/MapModelDb.java
@@ -5,7 +5,7 @@ import dk.itu.models.OsmElement;
 import java.util.List;
 
 public class MapModelDb extends MapModel {
-    public MapModelDb(float minLon, float minLat, float maxLat, List<List<OsmElement>> layers) {
+    public MapModelDb(double minLon, double minLat, double maxLat, List<List<OsmElement>> layers) {
         super();
         this.minLon = minLon;
         this.minLat = minLat;
